# apps/next-frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ticketeate API",
    "version": "1.0.0",
    "description": "Documentación de la API de Ticketeate (OAS 3.0)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Desarrollo local" }
  ],
...
  "paths": {
...
    "/api/cupones": {
      "get": {
        "tags": ["Cupones"],
        "summary": "Obtener cupones (última versión por cupón) de un evento",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID del evento"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cupones": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "cuponid": { "type": "string" },
                          "eventoid": { "type": "string" },
                          "codigo": { "type": "string" },
                          "porcentaje_descuento": { "type": "number" },
                          "fecha_creacion": { "type": "string", "format": "date-time" },
                          "fecha_expiracion": { "type": "string", "format": "date-time" },
                          "limite_usos": { "type": "integer" },
                          "usos_actuales": { "type": "integer" },
                          "estado": { "type": "string" },
                          "version": { "type": "integer" },
                          "changed_at": { "type": "string", "format": "date-time" },
                          "changed_by": { "type": "string" },
                          "change_type": { "type": "string" }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "cupones": [
                    {
                      "cuponid": "cup_123",
                      "eventoid": "evt_001",
                      "codigo": "EARLYBIRD",
                      "porcentaje_descuento": 20,
                      "fecha_creacion": "2025-10-01T10:00:00.000Z",
                      "fecha_expiracion": "2025-12-31T23:59:59.000Z",
                      "limite_usos": 100,
                      "usos_actuales": 5,
                      "estado": "ACTIVO",
                      "version": 3,
                      "changed_at": "2025-11-08T13:00:00.000Z",
                      "changed_by": "usr_abc",
                      "change_type": "UPDATE"
                    }
                  ]
                }
              }
            }
          },
          "400": { "description": "eventId es requerido" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Sin permisos" },
          "500": { "description": "Error del servidor" }
        }
      },
      "post": {
        "tags": ["Cupones"],
        "summary": "Crear un nuevo cupón para un evento",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventId","codigo","porcentaje_descuento","fecha_expiracion","limite_usos"],
                "properties": {
                  "eventId": { "type": "string" },
                  "codigo": { "type": "string" },
                  "porcentaje_descuento": { "type": "number" },
                  "fecha_expiracion": { "type": "string", "format": "date-time" },
                  "limite_usos": { "type": "integer" }
                }
              },
              "example": {
                "eventId": "evt_001",
                "codigo": "EARLY2025",
                "porcentaje_descuento": 15,
                "fecha_expiracion": "2025-12-31T23:59:59.000Z",
                "limite_usos": 200
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "cupon": { "type": "object" } }
                },
                "example": {
                  "cupon": {
                    "cuponid": "cup_456",
                    "eventoid": "evt_001",
                    "codigo": "EARLY2025",
                    "porcentaje_descuento": 15,
                    "fecha_creacion": "2025-11-08T14:10:00.000Z",
                    "fecha_expiracion": "2025-12-31T23:59:59.000Z",
                    "limite_usos": 200,
                    "usos_actuales": 0,
                    "estado": "ACTIVO",
                    "version": 1,
                    "updated_by": "usr_abc"
                  }
                }
              }
            }
          },
          "400": { "description": "Faltan campos o duplicado" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Sin permisos" },
          "500": { "description": "Error del servidor" }
        }
      },
      "patch": {
        "tags": ["Cupones"],
        "summary": "Actualizar un cupón (append-only history)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["cuponId","eventId"],
                "properties": {
                  "cuponId": { "type": "string" },
                  "eventId": { "type": "string" },
                  "codigo": { "type": "string" },
                  "porcentaje_descuento": { "type": "number" },
                  "fecha_expiracion": { "type": "string", "format": "date-time" },
                  "limite_usos": { "type": "integer" },
                  "estado": { "type": "string" }
                }
              },
              "example": {
                "cuponId": "cup_456",
                "eventId": "evt_001",
                "codigo": "EARLY2025-2",
                "porcentaje_descuento": 18,
                "fecha_expiracion": "2026-01-31T23:59:59.000Z",
                "limite_usos": 250,
                "estado": "ACTIVO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "cupon": { "type": "object" } } },
                "example": {
                  "cupon": {
                    "cuponid": "cup_456",
                    "eventoid": "evt_001",
                    "codigo": "EARLY2025-2",
                    "porcentaje_descuento": 18,
                    "fecha_creacion": "2025-11-08T14:10:00.000Z",
                    "fecha_expiracion": "2026-01-31T23:59:59.000Z",
                    "limite_usos": 250,
                    "usos_actuales": 0,
                    "estado": "ACTIVO",
                    "version": 2,
                    "changed_by": "usr_abc",
                    "change_type": "UPDATE"
                  }
                }
              }
            }
          },
          "400": { "description": "cuponId/eventId requeridos o conflicto" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Sin permisos" },
          "404": { "description": "No encontrado" },
          "500": { "description": "Error del servidor" }
        }
      },
      "delete": {
        "tags": ["Cupones"],
        "summary": "Eliminar (soft delete) un cupón",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "cuponId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "eventId", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Eliminado",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "cupon": { "type": "object" } } },
                "example": {
                  "cupon": {
                    "cuponid": "cup_456",
                    "eventoid": "evt_001",
                    "codigo": "EARLY2025-2",
                    "porcentaje_descuento": 18,
                    "fecha_creacion": "2025-11-08T14:10:00.000Z",
                    "fecha_expiracion": "2026-01-31T23:59:59.000Z",
                    "limite_usos": 250,
                    "usos_actuales": 0,
                    "estado": "INACTIVO",
                    "version": 3,
                    "changed_by": "usr_abc",
                    "change_type": "DELETE"
                  }
                }
              }
            }
          },
          "400": { "description": "cuponId y eventId requeridos" },
          "401": { "description": "No autorizado" },
          "403": { "description": "Sin permisos" },
          "404": { "description": "No encontrado" },
          "500": { "description": "Error del servidor" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}